fixing #187

+++ lib/msprime.h
@@ -900,7 +900,7 @@ void migration_table_print_state(migration_table_t *self, FILE *out);
 const char * msp_strerror(int err);
 void __msp_safe_free(void **ptr);
 
-double compute_falling_factorial_log(double m);
+double compute_falling_factorial_log(unsigned int  m);
 
 #define msp_safe_free(pointer) __msp_safe_free((void **) &(pointer))
